requirements.txt
django~=2.2.0
Pillow
django-debug-toolbar
